+++ Modules/Intent.Modules.AspNetCore.Swashbuckle/NugetPackages.cs
@@ -19,10 +19,10 @@ namespace Intent.Modules.AspNetCore.Swashbuckle
                 (framework) => framework switch
                     {
                         ( >= 2, 0) => new PackageVersion("7.0.0")
+                            .WithNugetDependency("Microsoft.Extensions.ApiDescription.Server", "6.0.5")
                             .WithNugetDependency("Swashbuckle.AspNetCore.Swagger", "7.0.0")
                             .WithNugetDependency("Swashbuckle.AspNetCore.SwaggerGen", "7.0.0")
-                            .WithNugetDependency("Swashbuckle.AspNetCore.SwaggerUI", "7.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.ApiDescription.Server", "6.0.5"),
+                            .WithNugetDependency("Swashbuckle.AspNetCore.SwaggerUI", "7.0.0"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{SwashbuckleAspNetCorePackageName}'"),
                     }
                 );